requirements.txt
apache-airflow==2.7.3
apache-airflow-providers-postgres
apache-airflow-providers-http==4.6.0
dbt-core
dbt-postgres
psycopg2-binary